+++ hardware/YouLess.h
@@ -38,6 +38,6 @@ class CYouLess : public CDomoticzHardwareBase
 	bool m_bHaveP1OrS0;
 	P1Power m_p1power;
 	P1Gas m_p1gas;
-	unsigned long m_lastgasusage;
+	uint32_t m_lastgasusage;
 	time_t m_lastSharedSendGas;
 };
